requirements.txt
pyyaml
mujoco-py      # or mujoco>=2.3
gymnasium
stable-baselines3
torch
unitree_legged_sdk
numpy
